refactor: relocate spec and widgetbook harnesses outside support tree (#4117)
Move widgetbook, combat-ui, move-dialogs, and dialogue-overlays spec
harness modules to app/test/ with shrink-only re-export stubs; ratchet
appTestSupportLocCeiling 289 → 164 (−125 LOC).

Refs #4117
