+++ app/test/dialogue_overlays_specs_test_support_test.dart
@@ -19,7 +19,7 @@ void main() {
 
   test('dialogueOverlaysLibraryUnitSource reads this support library', () {
     final source = dialogueOverlaysLibraryUnitSource(
-      'test/support/dialogue_overlays_specs_test_support.dart',
+      'test/dialogue_overlays_specs_test_support.dart',
     );
     expect(source, contains('wrapGameStartIntroOverlay'));
   });
