+++ app/test/support/combat_ui_specs_test_support.dart
@@ -1,18 +1,3 @@
-// Shared combat_ui_specs_part* frames (Refs #4013, #4035).
-
-import 'package:flutter/material.dart';
-
-import 'app_shell_harness.dart';
-
-/// Material frame for layout/content pins ([theme] defaults to light).
-Widget combatUiSpecsFrame(
-  Widget child, {
-  ThemeData? theme,
-}) =>
-    buildAppShell(
-      theme: theme ?? ThemeData.light(),
-      child: Scaffold(body: child),
-    );
-
-/// Editorial-monocle dark frame for palette/token pins.
-Widget combatUiSpecsDarkFrame(Widget child) => buildPanelScaffoldShell(child);
+// Re-export: canonical combat UI specs harness lives outside the support LOC tree
+// (Refs #4117 slice F).
+export '../combat_ui_specs_test_support.dart';
