+++ tool/check_app_test_support_loc.dart
@@ -15,7 +15,7 @@ const String appTestSupportRelativeDir = 'app/test/support';
 /// Post-#4035 densify support-tree ceiling (physical LOC).
 /// Ratcheted after wave-9 slice F panel/golden harness relocate outside support tree
 /// (Refs #4117).
-const int appTestSupportLocCeiling = 289;
+const int appTestSupportLocCeiling = 164;
 
 /// Counts physical lines of all `*.dart` files under [dir].
 int countAppTestSupportPhysicalLoc(Directory dir) {
